norm broken fix to make the flags www , segfault over one of the test in the main  sometimes

+++ srcs/parsing/ft_sequence.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_sequence.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tpacaud <tpacaud@student.42.fr>            +#+  +:+       +#+        */
+/*   By: fratardi <fratardi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/05 02:07:09 by tpacaud           #+#    #+#             */
-/*   Updated: 2019/08/13 06:05:55 by tpacaud          ###   ########.fr       */
+/*   Updated: 2019/08/15 07:46:05 by fratardi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int			ft_init_sequence(char *str, t_no_syntax *content, size_t len)
 	content->extra = 0;
 	content->left = 0;
 	i = 1;
-	while (ft_strchr("-+ 0#", str[i]) && str[i] && i < len)
+	while (ft_strchr("-+ 0#", str[i]) && str[i] && /**/(unsigned int)/**/i < len)
 	{
 		(str[i] == '-') ? (content->left = 1) : 0;
 		(str[i] == '0') ? (content->extra = 1) : 0;
@@ -28,7 +28,7 @@ int			ft_init_sequence(char *str, t_no_syntax *content, size_t len)
 	}
 	content->width = ft_atoi((str + i));
 	content->width -= (content->width) ? 1 : 0;
-	while (str[i] && (ft_isdigit(str[i]) || str[i] == '.') && i < len)
+	while (str[i] && (ft_isdigit(str[i]) || str[i] == '.') && /**/(unsigned int)/**/i < len)
 		i++;
 	return (i);
 }
